Reorder Wallet tests
Remove some redundant tests in Test_Wallet
Move Copper tests from Test_Wallet to UT_Wallet
