+++ Tests/IntegrationTests/GameServer/Test_Wallet.cs
@@ -23,76 +23,6 @@ namespace DOL.Integration.Gameserver
         }
 
         [Test]
-        public void GetBalance_OfCopper_Init_GetBalanceForCopperIsZero()
-        {
-            var wallet = CreateWallet();
-
-            var actual = wallet.GetBalance(Currency.Copper);
-            var expected = Currency.Copper.Mint(0);
-            Assert.That(actual, Is.EqualTo(expected));
-        }
-
-        [Test]
-        public void AddMoney_OneCopper_GetBalanceForCopperIsOne()
-        {
-            var wallet = CreateWallet();
-
-            wallet.AddMoney(Currency.Copper.Mint(1));
-
-            var actual = wallet.GetBalance(Currency.Copper);
-            var expected = Currency.Copper.Mint(1);
-            Assert.That(actual, Is.EqualTo(expected));
-        }
-
-        [Test]
-        public void RemoveMoney_OneCopper_GetBalanceForCopperIsZero()
-        {
-            var wallet = CreateWallet();
-
-            wallet.RemoveMoney(Currency.Copper.Mint(1));
-
-            var actual = wallet.GetBalance(Currency.Copper);
-            var expected = Currency.Copper.Mint(0);
-            Assert.That(actual, Is.EqualTo(expected));
-        }
-
-        [Test]
-        public void RemoveMoney_OneCopper_Init_False()
-        {
-            var wallet = CreateWallet();
-
-            var actual = wallet.RemoveMoney(Currency.Copper.Mint(1));
-
-            var expected = false;
-            Assert.That(actual, Is.EqualTo(expected));
-        }
-
-        [Test]
-        public void RemoveMoney_TwoCopper_AddOneCopperPrior_GetBalanceForCopperIsOne()
-        {
-            var wallet = CreateWallet();
-            wallet.AddMoney(Currency.Copper.Mint(1));
-
-            wallet.RemoveMoney(Currency.Copper.Mint(2));
-
-            var actual = wallet.GetBalance(Currency.Copper);
-            var expected = Currency.Copper.Mint(1);
-            Assert.That(actual, Is.EqualTo(expected));
-        }
-
-        [Test]
-        public void RemoveMoney_OneCopper_AddOneCopperPrior_True()
-        {
-            var wallet = CreateWallet();
-            wallet.AddMoney(Currency.Copper.Mint(1));
-
-            var actual = wallet.RemoveMoney(Currency.Copper.Mint(1));
-
-            var expected = true;
-            Assert.That(actual, Is.EqualTo(expected));
-        }
-
-        [Test]
         public void GetBalance_Aurulite_Init_Zero()
         {
             var wallet = CreateWallet();
